requirements.txt
numpy==2.*
pillow==10.*
opencv-python-headless==4.*
fonttools
fontbakery
gftools==0.9.*
git+https://github.com/googlefonts/diffenator2
shaperglot==0.6.*
dehinter==4.*

ufoLib2